git-svn-id: https://opentk.svn.sourceforge.net/svnroot/opentk@27 ebc5dd9b-fb1d-0410-b6f8-d24c324e9604

+++ Source/OpenGL/OpenGL/Wgl.cs
@@ -1,6 +1,8 @@
+#region License
 /* Copyright (c) 2006 Stephen Apostolopoulos
  * See license.txt for license info
  */
+#endregion
 
 using System;
 using System.Collections.Generic;
@@ -239,7 +241,7 @@ namespace OpenTK.OpenGL
         public static extern int ShareLists(int r1, int r2);
 
         [DllImport(_dll_name, EntryPoint = "wglGetProcAddress")]
-        public static extern int GetProcAddress(string funcname);
+        public static extern IntPtr GetProcAddress(string funcname);
 
         #endregion
     }
